+++ src/main/java/net/mohron/skyclaims/config/type/MiscConfig.java
@@ -8,12 +8,18 @@ import java.util.List;
 
 @ConfigSerializable
 public class MiscConfig {
-	@Setting(value = "Island-on-Join", comment = "Automatically create an island for a player on join")
-	private boolean islandOnJoin = false;
-	@Setting(value = "Create-Commands", comment = "Commands to run on island creation and reset")
-	private List<String> createCommands = new ArrayList<>();
-	@Setting(value = "Reset-Commands", comment = "Commands to run on island resets only")
-	private List<String> resetCommands = new ArrayList<>();
+	@Setting(value = "Island-on-Join", comment = "Automatically create an island for a player on join.")
+	private boolean islandOnJoin;
+	@Setting(value = "Create-Commands", comment = "Commands to run on island creation and reset.")
+	private List<String> createCommands;
+	@Setting(value = "Reset-Commands", comment = "Commands to run on island resets only.")
+	private List<String> resetCommands;
+
+	public MiscConfig() {
+		islandOnJoin = false;
+		createCommands = new ArrayList<>();
+		resetCommands = new ArrayList<>();
+	}
 
 	public boolean createIslandOnJoin() {
 		return islandOnJoin;