+++ src/main/java/net/mohron/skyclaims/config/type/WorldConfig.java
@@ -10,11 +10,17 @@ import org.spongepowered.api.world.World;
 @ConfigSerializable
 public class WorldConfig {
 	@Setting(value = "SkyClaims-World", comment = "Name of the world to manage islands in. Default: world")
-	private String worldName = Sponge.getGame().getServer().getDefaultWorldName();
+	private String worldName;
 	@Setting(value = "Island-Height", comment = "Height to build islands at. Default: 72")
-	private int defaultHeight = 72;
+	private int defaultHeight;
 	@Setting(value = "Spawn-Regions", comment = "The height & width of regions to reserve for Spawn. Default: 1")
-	private int spawnRegions = 1;
+	private int spawnRegions;
+
+	public WorldConfig() {
+		worldName = Sponge.getGame().getServer().getDefaultWorldName();
+		defaultHeight = 72;
+		spawnRegions = 1;
+	}
 
 	public World getWorld() {
 		return SkyClaims.getInstance().getGame().getServer().getWorld(worldName).orElse(WorldUtil.getDefaultWorld());